scripts:Add exception for error 1790
VALIDATION_ERROR_01790 is a special case that provides an exception
for valid usage when VK_AMD_negative_viewport_height extension is
enabled. There's no actual error to flag here, but the code does
handle the exception case correctly so adding it as an exception
in the stat script as well to reflect that it's implemented.