scripts:Update unique error enums as of 11/29/16
This was a beast of an update. Reconfiguration in the spec
caused various section numbers to move which created a need to remap
a ton of ids.

Also a large bulk of new valid usage blocks for new extensions as
well as legitimate addition and subtraction of core valid usage
statements.

I'd focus some time on updating the script to handle these cases
better but now that Khronos has agreed to integrating the ids into the
spec itself, that will eliminate the need for the updates once the
integration is complete so I'll focus my efforts there going fwd and
hopefully only have to perform 1 or 2 more of these updates.

That resulted in the following cmd line to do the update:

python spec.py -update -remap 182-166,85:273-252,18:301-270,49:\
397-368,15:919-768:2076-1600,61:2340-1665,13:697-1719:2137-1794,30:\
2169-1824,7:2176-1834:2178-1831,3:2181-1836,4:2185-1843:2187-1840,3:\
2190-1844,39:2233-1883,40:2284-1923,39:2324-1962,16:2348-1978:\
2354-1979:2368-1980,7:2375-1988:2376-1990:2377-1998,24:2418-2043,12:\
2402-2022,15:2417-2040:167-2059:268-2067,5:291-2072,10:2355-1679,2:\
2357-1683,11:2229-2327,4:2401-2343:2273-2331,11:2323-2342